Criadas novas classes: "BackofficeGerenciaProdutoController" e "BackofficeVizualizaProdutoController" e validação de cadastro de produto pronto.

+++ ProjetoIntegrador/src/main/java/br/senac/pi4/ProjetoIntegrador/Service/ProdutoService.java
@@ -26,5 +26,5 @@ public interface ProdutoService extends Serializable {
     
     public void alterar(Produto produto);
     
-    public void remover(Integer codigoProduto);
+    public void remover(Long codigoProduto);
 }
